- added additional rules for site access - move contact info to separate class - move user principal to separate class
