+++ src/main/java/business/config/BusinessSecurityConfig.java
@@ -1,8 +1,6 @@
 package business.config;
 
-import config.AppInitializer;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationManager;
@@ -18,9 +16,6 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.access.AccessDeniedHandler;
 import org.springframework.security.web.context.AbstractSecurityWebApplicationInitializer;
 
-import java.util.Arrays;
-import java.util.List;
-
 @Configuration
 @EnableWebSecurity
 @EnableGlobalMethodSecurity(securedEnabled = true, prePostEnabled = true)
@@ -33,22 +28,8 @@ public class BusinessSecurityConfig extends WebSecurityConfigurerAdapter {
     private AccessDeniedHandler customAccessDeniedExceptionHandler;
 
     @Autowired
-    private ApplicationContext context;
-
-    @Autowired
     public void configureGlobal(AuthenticationManagerBuilder auth) throws Exception {
-        List<String> profiles = Arrays.asList(context.getEnvironment().getActiveProfiles());
-        if (profiles.contains(AppInitializer.PROD_PROFILE_MODE)) {
-            auth.authenticationProvider(authenticationProvider());
-        } else if (profiles.contains(AppInitializer.DEV_PROFILE_MODE)) {
-            auth.inMemoryAuthentication().passwordEncoder(getPasswordEncoder())
-                    .withUser("admin").password("nimda");
-        } else {
-            throw new RuntimeException(
-                    "'" + AppInitializer.PROD_PROFILE_MODE +
-                    "' or '" + AppInitializer.DEV_PROFILE_MODE  +
-                            "' profile should be an active");
-        }
+        auth.authenticationProvider(authenticationProvider());
     }
 
     @Bean
@@ -72,10 +53,15 @@ public class BusinessSecurityConfig extends WebSecurityConfigurerAdapter {
                 and().
                 formLogin().
                 and().
+                httpBasic().
+                and().
+                logout().
+                and().
+                rememberMe().
+                and().
                 exceptionHandling().
                 accessDeniedHandler(customAccessDeniedExceptionHandler).
                 accessDeniedPage("/login");
-
     }
 
     @Bean(name = "passwordEncoder")
