+++ src/main/java/business/model/user/User.java
@@ -22,29 +22,6 @@ public @Data class User extends BaseModel {
     private String email;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "ROLE_ID")
-    private Role role;
-
-    @Column(nullable = false, unique = true, name = "LOGIN")
-    private String login;
-
-    private String password;
-
-    @Transient
-    private String confirmPassword;
-
-    private boolean enabled;
-
-    @Column(name = "account_non_expired")
-    private boolean accountNonExpired;
-
-    @Column(name = "account_non_locked")
-    private boolean accountNonLocked;
-
-    @Column(name = "credentials_non_expired")
-    private boolean credentialsNonExpired;
-
-    @ManyToOne(fetch = FetchType.LAZY)
     @JoinTable(
             name = "USER_COMPANY",
             joinColumns =
@@ -53,4 +30,6 @@ public @Data class User extends BaseModel {
                 @JoinColumn(nullable = false)
     )
     private Company company;
+
+    private UserPrincipal principal;
 }
