+++ src/main/webapp/WEB-INF/views/index.jsp
@@ -1,6 +1,7 @@
 <html>
 <body>
 <h2>Welcome</h2>
+
 <a href="/project/list">Project list</a>
 <a href="/action">Action</a>
 <sec></sec>
